requirements.txt
Flask==0.9
Flask-SQLAlchemy==0.16
Jinja2==2.7
Mako==0.8.1
MarkupSafe==0.18
PIL==1.1.7
SQLAlchemy==0.8.1
Unidecode==0.04.13
Werkzeug==0.8.3
alembic==0.5.0
anyjson==0.3.3
argparse==1.2.1
certifi==0.0.8
chardet==2.1.1
distribute==0.6.45
dnspython==1.11.0
facebook-sdk==0.4.0
gunicorn==17.5
indextank==1.0.6
ipython==1.0.0
psycopg2==2.5.1
python-dateutil==2.1
redis==2.7.6
requests==1.2.3
simplejson==3.3.0
six==1.3.0
wsgiref==0.1.2
xmlrpclib==1.0.1
